requirements.txt
streamlit
fpdf
requests
python-dotenv